refactor(traces): split the IO toolbar into one file per concern
The new `IOViewerToolbar.tsx` held five things: the presentational action
buttons, the action model and its permission gates, the format options, the
overflow measurement, and the toolbar itself. That is 665 lines in one file,
over the repo's 300-line limit for a new file.

Each concern now has its own module, and the toolbar file keeps only the row
it measures:

- `IOToolbarButtons.tsx` — the buttons an action renders.
- `ioActions.tsx` — the `IOAction` model, the gates, and `useIOActions`.
- `ioFormatOptions.ts` — the format options and their submode toggles.
- `IOViewerToolbar.tsx` — the action slot, the measured row, the toolbar.

`markdownSubmode` now reads the exported `MarkdownSubmode` instead of writing
`"rendered" | "source"` again in four places, so a new member cannot narrow
this prop against the state hook that supplies it.

The attribute-value test helper takes named parameters.
