+++ platform/app/src/features/traces-v2/components/TraceDrawer/IOViewerToolbar.tsx
@@ -1,32 +1,15 @@
-import { Box, Button, Flex, HStack, Icon } from "@chakra-ui/react";
-import { forwardRef, useMemo, useRef, useState } from "react";
-import type { IconType } from "react-icons";
-import {
-  LuCheck,
-  LuCode,
-  LuCopy,
-  LuEye,
-  LuLanguages,
-  LuLightbulb,
-  LuList,
-  LuMessageSquare,
-  LuPlay,
-} from "react-icons/lu";
-import { useOrganizationTeamProject } from "~/hooks/useOrganizationTeamProject";
-import { useGoToSpanInPlaygroundTabUrlBuilder } from "~/prompts/prompt-playground/hooks/useLoadSpanIntoPromptPlayground";
-import {
-  type TraceAnchor,
-  useAnchoredAnnotations,
-} from "../../hooks/useAnchoredAnnotations";
-import { useCopyToClipboard } from "../../hooks/useCopyToClipboard";
+import { Box, Flex, HStack, Icon } from "@chakra-ui/react";
+import { useMemo, useRef } from "react";
+import type { TraceAnchor } from "../../hooks/useAnchoredAnnotations";
 import { useOverflowVisibility } from "../../hooks/useOverflowVisibility";
 import type { useTextTranslation } from "../../hooks/useTextTranslation";
 import { OverflowMenu } from "../shared/OverflowMenu";
-import { FieldCommentButton } from "./anchoredComments/FieldCommentButton";
-import { AnnotationPopover } from "./conversationView/AnnotationPopover";
 import { FormatSelect } from "./FormatSelect";
+import { CopyButton } from "./IOToolbarButtons";
+import { type IOAction, useIOActions } from "./ioActions";
+import { formatSelectOptions } from "./ioFormatOptions";
 import type { ChatLayout } from "./transcript";
-import type { ViewFormat } from "./useIOViewerState";
+import type { MarkdownSubmode, ViewFormat } from "./useIOViewerState";
 
 /**
  * Room held back on the right edge of the actions row for the overflow
@@ -34,411 +17,6 @@ import type { ViewFormat } from "./useIOViewerState";
  * viz tab strip uses for the same trigger.
  */
 const OVERFLOW_TRIGGER_RESERVE_PX = 26;
-
-/**
- * The shape every toolbar action wears: icon, label, and the row's height.
- *
- * It stops the click from travelling, because the panel header behind it
- * toggles the panel collapsed.
- */
-const ActionButton = forwardRef<
-  HTMLButtonElement,
-  {
-    icon: IconType;
-    label: string;
-  } & React.ComponentProps<typeof Button>
->(function ActionButton({ icon, label, ...buttonProps }, ref) {
-  return (
-    <Button
-      ref={ref}
-      size="xs"
-      variant="ghost"
-      color="fg.muted"
-      gap={1.5}
-      paddingX={2}
-      height="22px"
-      onClick={(e) => e.stopPropagation()}
-      {...buttonProps}
-    >
-      <Icon as={icon} boxSize={3} />
-      {label}
-    </Button>
-  );
-});
-
-/**
- * Opens this span in the prompt playground. A real anchor, so the browser's
- * own gestures for a new tab or window still work on it.
- */
-function PlaygroundButton({ href }: { href: string }) {
-  return (
-    <Button
-      asChild
-      size="xs"
-      variant="ghost"
-      color="fg.muted"
-      gap={1.5}
-      paddingX={2}
-      height="22px"
-    >
-      <a
-        href={href}
-        target="_blank"
-        rel="noreferrer noopener"
-        onClick={(e) => e.stopPropagation()}
-      >
-        <Icon as={LuPlay} boxSize={3} />
-        Open in Playground
-      </a>
-    </Button>
-  );
-}
-
-/** Opens the annotation composer in `suggest` mode against this field. */
-function SuggestCorrectionButton({
-  traceId,
-  output,
-  anchor,
-}: {
-  traceId: string;
-  output: string;
-  /** The field the suggestion corrects. */
-  anchor: TraceAnchor;
-}) {
-  const [open, setOpen] = useState(false);
-  return (
-    <AnnotationPopover
-      traceId={traceId}
-      output={output}
-      mode="suggest"
-      anchorKind={anchor.anchorKind}
-      anchorId={anchor.anchorId}
-      anchorPath={anchor.anchorPath}
-      open={open}
-      onOpenChange={setOpen}
-      trigger={<ActionButton icon={LuLightbulb} label="Suggest edit" />}
-    />
-  );
-}
-
-/**
- * Turns the translated view on and off. The label is the whole state: what it
- * offers next while idle, and what it is doing while the request is out.
- */
-function TranslateButton({
-  isActive,
-  isLoading,
-  onToggle,
-}: {
-  isActive: boolean;
-  isLoading: boolean;
-  onToggle: () => void;
-}) {
-  return (
-    <ActionButton
-      icon={LuLanguages}
-      label={
-        isLoading ? "Translating…" : isActive ? "Show original" : "Translate"
-      }
-      aria-pressed={isActive}
-      color={isActive ? "blue.fg" : "fg.muted"}
-      disabled={isLoading}
-      onClick={(e) => {
-        e.stopPropagation();
-        onToggle();
-      }}
-    />
-  );
-}
-
-/**
- * Copies what the panel currently displays, and says so with a check for a
- * moment. Icon only: it is the last control in the row and needs no width.
- */
-function CopyButton({ text }: { text: string }) {
-  const { copied, copy } = useCopyToClipboard();
-
-  const handleCopy = (e: React.MouseEvent) => {
-    e.stopPropagation();
-    copy(text);
-  };
-
-  return (
-    <Button
-      size="xs"
-      variant="ghost"
-      onClick={handleCopy}
-      aria-label="Copy to clipboard"
-      padding={0}
-      minWidth="auto"
-      height="auto"
-    >
-      <Icon
-        as={copied ? LuCheck : LuCopy}
-        boxSize={3}
-        color={copied ? "green.fg" : "fg.subtle"}
-      />
-    </Button>
-  );
-}
-
-/** One toolbar action: its inline rendering and its overflow-menu row. */
-type IOAction = {
-  id: string;
-  menuLabel: string;
-  menuIcon: IconType;
-  disabled?: boolean;
-  render: () => React.ReactNode;
-};
-
-/**
- * The toolbar actions a panel offers, in render order.
- *
- * Each action keeps its real component mounted even while collapsed, so
- * popover wiring and permission gates live in one place and picking the
- * action out of the overflow menu clicks the same control.
- */
-function buildIOActions({
-  translation,
-  traceId,
-  fieldAnchor,
-  originalContent,
-  showComment,
-  showSuggest,
-  playgroundHref,
-}: {
-  translation: ReturnType<typeof useTextTranslation>;
-  traceId: string | undefined;
-  fieldAnchor: TraceAnchor | null;
-  originalContent: string;
-  showComment: boolean;
-  showSuggest: boolean;
-  playgroundHref: string;
-}): IOAction[] {
-  const actions: IOAction[] = [
-    {
-      id: "translate",
-      menuLabel: translation.isLoading
-        ? "Translating…"
-        : translation.isActive
-          ? "Show original"
-          : "Translate",
-      menuIcon: LuLanguages,
-      disabled: translation.isLoading,
-      render: () => (
-        <TranslateButton
-          isActive={translation.isActive}
-          isLoading={translation.isLoading}
-          onToggle={translation.toggle}
-        />
-      ),
-    },
-  ];
-
-  if (traceId && fieldAnchor && showComment) {
-    actions.push({
-      id: "comment",
-      menuLabel: "Comment",
-      menuIcon: LuMessageSquare,
-      render: () => (
-        <FieldCommentButton traceId={traceId} anchor={fieldAnchor} />
-      ),
-    });
-  }
-
-  if (traceId && fieldAnchor && showSuggest) {
-    actions.push({
-      id: "suggest",
-      menuLabel: "Suggest edit",
-      menuIcon: LuLightbulb,
-      // Every field this viewer shows is one a correction can replace, the
-      // trace's own input included. Corrections must be stored against the
-      // REAL text, never the translated variant the viewer happens to show.
-      render: () => (
-        <SuggestCorrectionButton
-          traceId={traceId}
-          output={originalContent}
-          anchor={fieldAnchor}
-        />
-      ),
-    });
-  }
-
-  if (playgroundHref) {
-    actions.push({
-      id: "playground",
-      menuLabel: "Open in Playground",
-      menuIcon: LuPlay,
-      render: () => <PlaygroundButton href={playgroundHref} />,
-    });
-  }
-
-  return actions;
-}
-
-/**
- * The format options for a panel, with the inline submode toggles the
- * active format carries.
- *
- * Both chat layouts (thread / bubbles) are available for any chat-shaped
- * content — input *or* output. Even with a single assistant reply, the
- * operator may want the bubble visual; conversely, a multi-message output
- * (rare but possible) benefits from the flat stack.
- */
-function formatSelectOptions({
-  formatOptions,
-  isChat,
-  chatLayout,
-  onChatLayoutChange,
-  markdownSubmode,
-  onMarkdownSubmodeChange,
-}: {
-  formatOptions: readonly ViewFormat[];
-  isChat: boolean;
-  chatLayout: ChatLayout;
-  onChatLayoutChange: (layout: ChatLayout) => void;
-  markdownSubmode: "rendered" | "source";
-  onMarkdownSubmodeChange: (submode: "rendered" | "source") => void;
-}) {
-  return formatOptions.map((option) => {
-    if (option === "pretty" && isChat) {
-      return {
-        value: option,
-        submodes: {
-          value: chatLayout,
-          onChange: onChatLayoutChange,
-          options: [
-            {
-              value: "thread",
-              label: "Thread",
-              icon: LuList,
-              tooltip: "Thread layout",
-            },
-            {
-              value: "bubbles",
-              label: "Bubbles",
-              icon: LuMessageSquare,
-              tooltip: "Bubble layout",
-            },
-          ],
-        },
-      };
-    }
-    if (option === "markdown") {
-      return {
-        value: option,
-        submodes: {
-          value: markdownSubmode,
-          onChange: onMarkdownSubmodeChange,
-          options: [
-            {
-              value: "rendered",
-              label: "Rendered",
-              icon: LuEye,
-              tooltip: "Rendered markdown view",
-            },
-            {
-              value: "source",
-              label: "Source",
-              icon: LuCode,
-              tooltip: "Source markdown view",
-            },
-          ],
-        },
-      };
-    }
-    return { value: option };
-  });
-}
-
-/**
- * Which annotation actions this reader gets, and where "Open in Playground"
- * points.
- *
- * The annotation gate mirrors AnchorCommentButton's own: writers always get
- * the action, readers only when there is something to read.
- */
-function useIOActionGates({
-  fieldAnchor,
-  spanId,
-  spanType,
-  mode,
-}: {
-  fieldAnchor: TraceAnchor | null;
-  spanId: string | undefined;
-  spanType: string | undefined;
-  mode: "input" | "output";
-}) {
-  const { hasPermission } = useOrganizationTeamProject();
-  const annotations = useAnchoredAnnotations();
-  const { buildUrl } = useGoToSpanInPlaygroundTabUrlBuilder();
-
-  const canAnnotate = hasPermission("annotations:manage");
-  return {
-    showComment:
-      fieldAnchor !== null &&
-      (canAnnotate || annotations.commentsAt(fieldAnchor).length > 0),
-    showSuggest: fieldAnchor !== null && canAnnotate,
-    // No explicit playground action — the loader auto-detects: opens the
-    // existing managed prompt at the traced version when one is linked,
-    // creates a fresh tab when not. One button, smart default.
-    playgroundHref:
-      spanType === "llm" && spanId && mode === "input"
-        ? (buildUrl(spanId)?.toString() ?? "")
-        : "",
-  };
-}
-
-/**
- * The panel's actions, with the annotation gates and the playground target
- * already applied.
- */
-function useIOActions({
-  translation,
-  traceId,
-  spanId,
-  spanType,
-  mode,
-  fieldAnchor,
-  originalContent,
-}: {
-  translation: ReturnType<typeof useTextTranslation>;
-  traceId: string | undefined;
-  spanId: string | undefined;
-  spanType: string | undefined;
-  mode: "input" | "output";
-  fieldAnchor: TraceAnchor | null;
-  originalContent: string;
-}): IOAction[] {
-  const { showComment, showSuggest, playgroundHref } = useIOActionGates({
-    fieldAnchor,
-    spanId,
-    spanType,
-    mode,
-  });
-
-  return useMemo(
-    () =>
-      buildIOActions({
-        translation,
-        traceId,
-        fieldAnchor,
-        originalContent,
-        showComment,
-        showSuggest,
-        playgroundHref,
-      }),
-    [
-      translation,
-      traceId,
-      fieldAnchor,
-      originalContent,
-      showComment,
-      showSuggest,
-      playgroundHref,
-    ],
-  );
-}
 
 /**
  * One action in the row. Collapsed, it keeps its real control mounted at zero
@@ -580,8 +158,8 @@ interface IOViewerToolbarProps {
   isChat: boolean;
   chatLayout: ChatLayout;
   onChatLayoutChange: (layout: ChatLayout) => void;
-  markdownSubmode: "rendered" | "source";
-  onMarkdownSubmodeChange: (submode: "rendered" | "source") => void;
+  markdownSubmode: MarkdownSubmode;
+  onMarkdownSubmodeChange: (submode: MarkdownSubmode) => void;
   translation: ReturnType<typeof useTextTranslation>;
   traceId: string | undefined;
   spanId: string | undefined;
